requirements.txt
pymongo>=4.3.3
mysql-connector-python>=8.0.33
PyYAML>=6.0
click>=8.1.3
tqdm>=4.65.0
colorama>=0.4.6
python-dotenv>=1.0.0